Added many tests, updated report, corrected bugs

+++ calcul.h
@@ -23,7 +23,7 @@ pos_t compute_position(graph_t* G,int idm,double t);
 pos_t compute_pos_intercep(graph_t* G, double alpha, pos_t pos, double t);
 
 /*Calcul le temps requis pour intercepter le mobile idm A PARTIR de la date t  et de la position pos*/
-double compute_interception(graph_t* G, pos_t* pos, int idm,int t, double* alpha);
+double compute_interception(graph_t* G, pos_t* pos, int idm,double t, double* alpha);
 
 /*Calcul la solution  de l'équation a*cos(x)+b*sin(x) = c pour x inconnu*/
 double compute_alpha(double a, double b, double c);
